481 rotation of output images in prostate mri anatomy (#482)
### Description
481-Rotation of output images in prostate_mri_anatomy

- The transformations applied to the test data during preprocessing are
not reversed during postprocessing. Thus, the resulting predictions have
a different rotation compared to the original images and labels. I added
an 'invertd()' function that reverses all the traceable transformations
in the preprocessing.
- Running the inference on a PC without a GPU leads to an error because
the handlers function 'CheckpointLoader' attempts to load the model onto
the GPU. I set the map_location argument to torch.device(device) so that
it automatically loads the model onto the CPU when a GPU is not
available.
- Additionally, I changed some of the config directories as they were
referring to the validation dataset rather than the test dataset.

### Status
Ready

### Please ensure all the checkboxes:
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Codeformat tests passed locally by running `./runtests.sh
--codeformat`.
- [x] Update `version` and `changelog` in `metadata.json` if changing an
existing bundle.
- [x] Please ensure the naming rules in config files meet our
requirements (please refer to: `CONTRIBUTING.md`).
- [x] Ensure versions of packages such as `monai`, `pytorch` and `numpy`
are correct in `metadata.json`.
- [x] Descriptions should be consistent with the content, such as
`eval_metrics` of the provided weights and TorchScript modules.
- [x] Files larger than 25MB are excluded and replaced by providing
download links in `large_file.yml`.
- [x] Avoid using path that contains personal information within config
files (such as use `/home/your_name/` for `"bundle_root"`).

---------

Signed-off-by: FaresAlMohamad <141377568+FaresAlMohamad@users.noreply.github.com>